Move RaindexVersionValidator to components-ui
